Rename cron -> periodic and add docs
