+++ README.md
@@ -54,6 +54,7 @@ impl Worker<PaymentReportArgs> for PaymentReportWorker {
 }
 ```
 
+
 ## Creating a Job
 
 There are several ways to insert a job, but for this example, we'll keep it simple. Given some worker, insert using strongly
@@ -99,6 +100,7 @@ sidekiq::perform_async(
 
 See more examples in `examples/demo.rs`.
 
+
 ## Starting the Server
 
 Below is an example of how you should create a `Processor`, register workers, include any
@@ -126,6 +128,38 @@ p.using(FilterExpiredUsersMiddleware::new(logger.clone()))
 // Start the server
 p.run().await;
 ```
+
+
+## Periodic Jobs
+
+Periodic cron jobs are supported out of the box. All you need to specify is a valid
+cron string and a worker instance. You can optionally supply arguments, a queue, a
+retry flag, and a name that will be logged when a worker is submitted.
+
+Example:
+
+```rust
+// Clear out all periodic jobs and their schedules
+periodic::destroy_all(redis).await?;
+
+// Add a new periodic job
+periodic::builder("0 0 8 * * *")?
+    .name("Email clients with an oustanding balance daily at 8am UTC")
+    .queue("reminders")
+    .args(EmailReminderArgs {
+        report_type: "outstanding_balance",
+    })?
+    .register(&mut p, EmailReminderWorker)
+    .await?;
+```
+
+The implementation relies on a sorted set in memory. It stores a json payload of the
+periodic job with a score set to the next scheduled UTC time of the cron string. All
+clients will periodically poll for changes and atomically update the score to the new
+next scheduled UTC time of the cron string. The worker that successfully changes the
+score atomically will enqueue a new job. Processes that don't atomically update the 
+score will move on. This implementation means periodic jobs never leave redis.
+
 
 ## Server Middleware
 
@@ -191,6 +225,7 @@ impl ServerMiddleware for FilterExpiredUsersMiddleware {
     }
 }
 ```
+
 
 ## License
